agentcore/requirements.txt
bedrock-agentcore
strands-agents
flask
boto3